Prodigy_SD_Task1_Temperature Converter
This is a simple program that converts temperatures between Celsius, Fahrenheit, and Kelvin scales. This program will prompt the user to input a temperature value and the original unit of measurement. This will then convert the temperature to the other two units and display the converted values to the user in the chosen scale according to the user need.