Create Prodigy_SD_Task2_Gussing Game
The aim of this project is to build a program that generates a random number and challenges the user to guess it. The program will prompt the user to input their guess, compare it to the generated number, and provide feedback if the guess is too high or too low. This continues until the user correctly guesses the number and then display the number of attempts it took to win the game.