Prodigy_SD_Task3_Contact Management
This is a simple program that allows users to store and manage contact information. The program will provide options to add a new contact by entering their name, phone number, and email address. It will also allow users to view their contact list, edit existing contacts, and delete contacts if needed. The program should store the contacts in memory or in a file for persistent storage.